refactor(prompt): extract system prompt to src/prompt.ts
Move buildDefaultSystemPrompt from src/index.ts to dedicated
src/prompt.ts module. Re-export from index for public API.
